move FBA and FBS schema dictionaries to own py file
