+++ flowsa/data_source_scripts/BLS_QCEW.py
@@ -15,8 +15,8 @@ import zipfile
 import io
 import pandas as pd
 import numpy as np
-from flowsa.common import US_FIPS, fba_default_grouping_fields, flow_by_activity_wsec_fields, \
-    flow_by_activity_mapped_wsec_fields
+from flowsa.common import US_FIPS, fba_default_grouping_fields
+from flowsa.schema import flow_by_activity_wsec_fields, flow_by_activity_mapped_wsec_fields
 from flowsa.flowbyfunctions import assign_fips_location_system, \
     aggregator
 from flowsa.dataclean import add_missing_flow_by_fields, replace_strings_with_NoneType
